Store relevant Owner metadata for MCG objects
Allow customers to keep track of object owner of an object via the S3 API

Currently when using boto3.client.list_objects, the metadata in the response under the "Owner" property is always set to
{'DisplayName': 'NooBaa', 'ID': '123'}
regardless of the name of the account that created the bucket.

Currently when using boto3.client.list_objects, the metadata in the response under the "Owner" property is always set to
{'DisplayName': 'NooBaa', 'ID': '123'}

regardless of the name of the account that created the bucket.

In Scope
Assign obc name or access key as the owner

Yes. document the approach we take
No

As an administrator, I should be able to check which account uploaded a given object to an MCG bucket.

Since ACLs are not suported, we are assigning obc owner as the owner of the object

Fixes: https://issues.redhat.com/browse/RHSTOR-5501

Signed-off-by: Vinayakswami Hariharmath <vharihar@redhat.com>